Rename three test classes so Surefire collects them
BeanMatcherCustomMatchOnInheritedFields, BeanMatcherIgnoreOrder and
JsonMatcherIgnoreOrder matched none of Surefire's default include patterns
(Test*, *Test, *Tests, *TestCase), are concrete, and have no subclass, so their
27 tests had never run. The Test suffix brings them in, in both of the
reflection-mode executions the core module runs.

The five inherited-fields tests and the 11 BeanMatcherIgnoreOrder tests pass
unchanged. The latter compare two Bean instances, so changes in serialisation
apply to both sides and cancel.

All 11 JsonMatcherIgnoreOrder expectations were stale and are regenerated from
current behaviour: null fields serialise by default, Set elements that serialise
identically are no longer collapsed, and object keys are emitted sorted rather
than in declaration order. The two nested-set cases therefore now assert three
elements where they asserted one. The class no longer pins
withoutSerializingNulls, so it runs on the default path like its bean twin.

Two nested-map expectations render each entry value-first. A complex-key map
entry serialises as a [key, value] array, that array is sorted like any other,
and every value in those fixtures carries a nested map whose "map": [ sorts
ahead of the key's "map": null. No code reads the position, so this affects only
how the file reads; both expectations carry a comment recording it.

Test-only. No approved file changes, so no CHANGELOG entry.
